Update config to have better sections (#85)
Allows you to disable any one of the outward connections, while also simplifying some of the config names and making it much easier to read and figure out what is happening.

I have tested this, but if someone else could test (preferably with the prod config) that would be great. I think I've caught all the changes.